Prototype for "Project" support
First of all: This is a prototype, so don't expect it to be finished ;-) 
With this version you should be able to specify various "projects" that you
would like to group your software into, and limit your search within the
selected projects. Did I mention that this is a prototype? ok.. The history
cache has been moved into $DATA_ROOT/historycache, and if you're using
Mercurial, you should probably pass the -H option to the indexer when you
generate the index database...