Create webapp as part of the normal build step
The webapp will also include the various configurations, so the user may
select the look'n'feel without rebuilding the application. By default the
webapp will use the look'n'feel (css-files and icons) located in the folder
named default. You may alter the content in this directory, or specify the
look'n'feel to use by using the -L option and create a configuration file
to OpenGrok.

ex:
java -jar opengrok.jar -R configuration.xml -W configuration.xml \
         -U localhost:2424 -L offwhite -n

will read the current configuration, set the look'n'feel to "offwhite" and
store the new configuration. The -U option I specified here was to update
the webapplicaton that listens to configuration changes on localhost:2424