OpenGrok (Shell Script) - Improve the readibility of recent additions to new users
- Changed usage to show 'index' has optional argument
- Factored out common code related to script location into a variable
- Renamed variables to assist new users (*)
- Regularised comment on newly added variables
- Better error handling for default logging properties installation
- In OpenGrokUsage() changed Progress to echo as it is not conditional text
- Added a few TODOs for possible future work

Possibly contentious changes (which will need more thought)
- Reverted to Non-Verbose operation
  (as I thought is was changed for debugging purposes)
- Script defaults versus opengrok.jar defaults need serious thought (**)

Footnote *: This script is first place most new users and installers
will have to customise, it should be as understandable as possible.  It
took me quite a while to convert run.sh into OpenGrok, as I was a new
user at that point, I'd like the way for all those that follow to be as
easy as possible, hence my continued tweaking of this script.

Footnote **:  I'd content that we should not have uncommented defaults
in this script that change opengrok.jar defaults, like "-r on" or "-m
100000" So I have a mixed bag in this commit, commenting out
REMOTE_REPOSITORIES, MAX_INDEXED_WORDS but leaving ENABLE_PROJECTS and
WEBAPP_CONFIG_ADDRESS enabled. This is deliberate to provoke a
discussion amongst the key developers of what they like the defaults to
be, and how this affects new users trying to understand and install
OpenGrok for the first time