Don't run diff jumper code when not viewing a diff.
The code that searches the document for difftable runs in every page
that loads utils.js. This makes loading of pages slower than necessary.

This patch factors out the diff jumper code from utils.js and puts it
in a separate file called diff.js which is only loaded by diff.jsp.