Clean up encoding and decoding of URL parameters.
mast.jsp needs to access some URL parameters in JavaScript code.
Currently, it does this in a Java snippet that escapes the parameter
value and prints it in the HTML. This patch changes it so that the
parameters are fetched by the JavaScript code itself, eliminating the
roundtrip via the Java and HTML domains, which simplifies the quoting.

Also remove Util.escapeXml(). Instead use Util.URIEncode(),
Util.htmlize() and Util.jsStringLiteral() as appropriate.
Util.escapeXml() was a hybrid of the latter two.

Also quote some values that were previously not quoted.