Fix JavaScript error in Internet Explorer.
#1063 introduced code that uses Array.map(). Internet Explorer
has supported the map() function since version 9, but OpenGrok
sends a meta header that makes Internet Explorer use IE8
compatibility mode, which causes the map() function to fail.

Since IE8 is no longer supported, except on a couple of embedded
platforms, let's try to disable the IE8 compatibility mode for now.
Hopefully, this shouldn't cause problems for anyone. If it turns out
that it does, we should revert this commit and rewrite the fix for
#1063 to not use the map() function.