Remove blank lines before the xml directive.
Most of the servlets add one or two blank lines before the xml
directive. The xml directive should be first in a well-formed xml
document.

This patch sets the "trim-directive-whitespaces" property in the
deployment descriptor, so that blank lines inserted by JSP directives
are removed. The deployment descriptor format was upgraded from 2.3
to 2.5 in order to support this property.