Simplify toggling of line numbers and annotations.
When line numbers or annotations are toggled, the anchors on every
line of the document are updated with a new class. It is easier to
just change the class of a parent element, such as the "body" element.

The motivation for this change was to speed up toggling of line
numbers in large documents. Unfortunately, it had no measurable
effect on the performance. It seems like most of the time is spent
doing the actual redrawing of the page, and the time spent on
manipulating the DOM is negligible in comparison.

It does however simplify the code that implements the toggling, and
less code is better.