CXref produces ill-formed XML when run on lex files.
JFlex files are analyzed with CAnalyzer. When CXref comes across a
JFlex rule that contains an escaped quote character, such as

  \" { yybegin(STRING); }

it starts a new string literal which is never terminated. The
resulting xref is not valid XML because it lacks an end tag.

This fix adds a rule to CXref which recognizes escaped quotes and
don't start a new string literal when it sees them.