Improve the validate-xref script.
The script currently wraps the xref contents in a <div> element before
it passes it to the validator. The actual xref generated by list.jsp
wraps it in <div><pre>...</pre></div>. Make the script do the same.
This removes some false positives, where the xref starts by closing
the <pre> tag and ends by opening a new one. It also makes the script
detect more errors, where the xref uses tags not allowed inside a
<pre> element.