Broken HTML generated by FortranXref
Special characters are not always escaped. (Since the Fortran
analyzer is used for .inc files, which is a pretty generic
extension, it is often used for files which are not really Fortran
files. The broken HTML was seen when using FortranXref to analyze
PHP files.)

Also, if the file does not have a final newline, the last <span>
element won't have an end tag. Use yypush()/yypop() to ensure all
<span> elements are closed.