Fix semantics of -H flag and disabling history generation
OpenGrok generates history regardless of what OPENGROK_GENERATE_HISTORY is set
to, for local repositories.  This variable currently only goes into effect for
remote repositories.  Change its semantics such that if -H is not passed, then
history generation and caching is disabled altogether for all repositories.
This is useful when indexing large local repositories such as a local mirror of
the AOSP sources worth 70GB. History generation in this case results in the
indexing never completing.

We also fix the semantics of OPENGROK_GENERATE_HISTORY in the wrapper script,
such that if its set to "off", then we don't pass -H at all to OpenGrok.

Signed-off-by: Joel Fernandes <agnel.joel@gmail.com>