Cure for missing file separator on subdirectory argument
After looking at the code there was a simple concatenation of a directory string with the data from the command line. A better technique would be to use the Paths interface from java.nio.file.Paths which will deal appropriately with file separation characters when they are missing or not.