Fix `READONLY_CONFIG_FILE` configurations not being merged
When `READONLY_CONFIG_FILE` is used, the file name passed to `merge_config_files` is the object as a string which is serialized and written into disk. This always results in a file named `<tempfile._TemporaryFileWrapper object at 0x7fa12bf82350>` on disk instead of the actual name and doesn't actually merge configs.