Add null guards
While the `@NotNull` decorator informs developers that the correct usage
is that the decorated variables should not be `null`, it does not enforce
such in any way.
This leads to cases where an incorrectly fed `null` could propagate far
away before being detected (such as via `NullPointerException`s).
This commit implements some `null` guards where needed for a fail-fast
treatment of `null` inputs.